Move dispatcher exceptions into exceptions module
We stopped short of this in #17, but I think since we're not going to
split out the dispatcher before 1.0 we should go ahead and do this for
the sake of more sensible documentation if nothing else.
