+++ aspen/exceptions.py
@@ -49,3 +49,19 @@ class TypecastError(Exception):
     def __init__(self, extension):
         self.msg = "Failure to typecast extension '{0}'".format(extension)
         Exception.__init__(self)
+
+
+class DispatchError(Exception):
+    def __init__(self, message):
+        self.message = message
+        Exception.__init__(self)
+
+class NotFound(DispatchError):
+    def __init__(self, message=''):
+        DispatchError.__init__(self, message if message else "not found")
+
+class AttemptedBreakout(NotFound): pass
+class UnindexedDirectory(NotFound): pass
+class Redirect(DispatchError): pass
+class RedirectFromIndexFilename(Redirect): pass
+class RedirectFromSlashless(Redirect): pass
